requirements.txt
fastapi
uvicorn
openai
python-dotenv
slowapi
tenacity
sqlalchemy
psycopg2-binary
pgvector
pytest
requests